Rename HostAllowList to host-allowlist
